pyproject.toml
[project]
name = "playstore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.3.0",
    "numpy>=1.23.0",
    "pandas>=1.5.0",
    "pytest>=7.0.0",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.2.0",
]
